pyproject.toml
[project]
name = "car-game"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "box2d>=2.3.10",
    "box2d-py>=2.3.8",
    "gymnasium>=1.2.1",
    "matplotlib>=3.10.7",
    "pillow>=10.4.0",
    "pygame-ce>=2.5.5",
    "pytest>=8.4.2",
    "swig>=4.3.1.post0",
]
